refactor(vibe3): break 11 module-internal circular dependencies
Breaks length-2 circular dependencies by replacing parent module imports
with direct source imports or relative imports.

Changes:
- runtime: heartbeat/cleanup/periodic_check executors import directly
  from models/config/orchestra.logging instead of vibe3.runtime
- orchestra: submodules import from models/domain/services instead
  of vibe3.orchestra parent module
- adapters: vibe_center uses relative import for registry access
- analysis: inspect_query_service uses relative import for dag_service

Impact:
- Reduced length-2 cycles: 15 → 4 (-73%)
- All tests passing (2505 tests collected)
- No functional changes, only import path refinements

Remaining: 4 deep architecture cycles (domain/events/services internal)
require protocol-based DI or event-driven refactoring.
