+++ src/vibe3/orchestra/issue_loader.py
@@ -6,7 +6,7 @@ from typing import TYPE_CHECKING
 
 from loguru import logger
 
-from vibe3.orchestra import OrchestraConfig
+from vibe3.models.orchestra_config import OrchestraConfig
 
 if TYPE_CHECKING:
     from vibe3.clients.github_client import GitHubClient
