+++ src/vibe3/orchestra/protocols.py
@@ -10,7 +10,9 @@ The definition below is maintained for backward compatibility during migration.
 
 from typing import Protocol
 
-from vibe3.orchestra import CheckResult, GitClient, IssueInfo
+from vibe3.clients.git_client import GitClient
+from vibe3.models.orchestration import IssueInfo
+from vibe3.services.check_service import CheckResult
 
 
 class CapacityServiceProtocol(Protocol):
