+++ src/vibe3/orchestra/queue_persistence_service.py
@@ -10,14 +10,11 @@ from typing import TYPE_CHECKING, Callable
 
 from loguru import logger
 
-from vibe3.orchestra import (
-    IssueInfo,
-    IssueState,
-    QueueEntry,
-    get_manager_usernames,
-    promote_progressed_entries,
-    should_skip_from_queue,
-)
+from vibe3.models.orchestration import IssueInfo, IssueState
+from vibe3.models.queue_entry import QueueEntry
+from vibe3.orchestra.queue_operations import promote_progressed_entries
+from vibe3.services.label_utils import should_skip_from_queue
+from vibe3.services.orchestra_helpers import get_manager_usernames
 
 if TYPE_CHECKING:
     from vibe3.clients.github_client import GitHubClient
