+++ src/vibe3/runtime/cleanup_executor.py
@@ -2,7 +2,8 @@
 
 from loguru import logger
 
-from vibe3.runtime import PeriodicCheckConfig, append_orchestra_event
+from vibe3.config.orchestra_config import PeriodicCheckConfig
+from vibe3.orchestra.logging import append_orchestra_event
 
 
 async def execute_expired_resource_cleanup(
